requirements.txt
streamlit
pandas
requests
python-dotenv
google-auth-oauthlib
openpyxl
